April 2025 data
SERTM2, a previously uncharacterized gene which was the focus of my classmate in GCD 3485 in fall 2023, has now also had a paper published on PubMed about it in March of 2025! Seemingly, a paper published in 2024 also focused on SERTM2, but did not refer to the gene by its official symbol in the paper title, and thus was not identified by this algorithm. This could present an argument for including paper abstracts in the analysis as well.